requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.4.1
Django==2.1.3
django-rest-framework==0.1.0
djangorestframework==3.9.0
idna==2.7
jwt==0.5.4
pycparser==2.19
pytz==2018.7
six==1.11.0
